+++ FishMarket/cart/templatetags/total_per_product.py
@@ -6,6 +6,6 @@ register = template.Library()
 @register.simple_tag()
 def total_per_product(price_per_kg, weight_per_unit, quantity):
     if not weight_per_unit:
-        return price_per_kg * quantity
+        return float(price_per_kg * quantity)
     else:
         return price_per_kg * weight_per_unit * quantity
